+++ src/bin/hook.ts
@@ -1,0 +1,5 @@
+#!/usr/bin/env bun
+
+import { launchTripwire } from './launch';
+
+launchTripwire('hook');